murmr/requirements.txt
faster-whisper==1.1.1
pynput==1.7.7
sounddevice==0.5.1
soundfile==0.13.1
pyperclip==1.9.0
pystray==0.19.5
Pillow==11.2.1
python-dotenv==1.1.0
numpy==2.2.4
requests>=2.28.0
